Convert id values from int to long in Java. It's already long-int in postgres, but java didn't know that.
git-svn-id: https://gd-ice.googlecode.com/svn/trunk/ice@597 d7922f76-c635-11de-8f3f-2f730f80caef

+++ src/main/java/org/jbei/ice/controllers/AccountController.java
@@ -20,7 +20,7 @@ import org.jbei.ice.lib.utils.Utils;
 import org.jbei.ice.web.PersistentSessionDataWrapper;
 
 public class AccountController {
-    public static Account get(int id) throws ControllerException {
+    public static Account get(long id) throws ControllerException {
         Account account = null;
 
         try {
